feat: strengthWeaknessFunnel에서 데이터 변환 로직 추가 및 데이터 전달 방식 개선
